% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_regression.R
\docType{class}
\name{AFTSurvivalRegressionModel-class}
\alias{AFTSurvivalRegressionModel-class}
\title{S4 class that represents a AFTSurvivalRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala AFTSurvivalRegressionWrapper}
}
\description{
S4 class that represents a AFTSurvivalRegressionModel
}
\note{
AFTSurvivalRegressionModel since 2.0.0
}
