% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{as_datetime.Column}
\alias{as_datetime.Column}
\title{Convert to datetime}
\usage{
as_datetime.Column(x, ...)
}
\arguments{
\item{x}{a Column object}

\item{...}{other argument(s), currently unused.}
}
\value{
a Column object
}
\description{
this function casts a Spark Column to timestamp type
}
