% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{avg}
\alias{avg}
\alias{avg,Column-method}
\title{avg}
\usage{
avg(x, ...)

\S4method{avg}{Column}(x)
}
\arguments{
\item{x}{Column to compute on or a GroupedData object.}

\item{...}{additional argument(s) when \code{x} is a GroupedData object.}
}
\description{
Aggregate function: returns the average of the values in a group.
}
\note{
avg since 1.4.0
}
\examples{
\dontrun{avg(df$c)}
}
\seealso{
Other aggregate functions: 
\code{\link{corr}()},
\code{\link{cov}()},
\code{\link{first}()},
\code{\link{last}()}
}
\concept{aggregate functions}
