% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\docType{class}
\name{BisectingKMeansModel-class}
\alias{BisectingKMeansModel-class}
\title{S4 class that represents a BisectingKMeansModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala BisectingKMeansModel}
}
\description{
S4 class that represents a BisectingKMeansModel
}
\note{
BisectingKMeansModel since 2.2.0
}
