% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{coalesce.Column}
\alias{coalesce.Column}
\title{Coalesce \code{Columns}}
\usage{
\method{coalesce}{Column}(...)
}
\arguments{
\item{...}{Column objects to be coalesces}
}
\value{
a Column object
}
\description{
Coalesces any number of Columns where precedence of the values is taken
as the order of the inputs.
}
