% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{coalesce}
\alias{coalesce}
\title{Reduce partitions OR Find first non-missing element}
\usage{
coalesce(...)
}
\arguments{
\item{...}{For the ETL case, this is the olumn or objects coercible to
Column to be coalesced. For the partition reducing case the first argument
should be a \code{spark_tbl} and the second should be an integer specifing
the number of partitions to reduce to.}
}
\description{
\code{coalesce} is used twice in Spark. The first use case is
to reduce the number of partitions of a Spark \code{DataFrame} without
a shuffle stage (in contrast to \code{repartition} which requires a shuffle).
The other use case is for ETL where it can be used on a Column object to
find the first non-missing element. See \code{?dplyr::coalese} for more info.
}
