% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\docType{methods}
\name{Column-functions}
\alias{Column-functions}
\alias{mean,Column-method}
\alias{xtfrm,Column-method}
\title{Column Functions}
\usage{
\S4method{mean}{Column}(x)

\S4method{xtfrm}{Column}(x)
}
\arguments{
\item{x}{a Column object}
}
\value{
a Column object
}
\description{
a collection of functions for Column objects
}
