% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{Column-missing}
\alias{Column-missing}
\alias{is.na,Column-method}
\alias{is.nan,Column-method}
\title{Check missing values in Column objects}
\usage{
\S4method{is.na}{Column}(x)

\S4method{is.nan}{Column}(x)
}
\arguments{
\item{x}{a Column object}
}
\description{
Check missing values in Column objects
}
