% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{Column-type}
\alias{Column-type}
\alias{as.character.Column}
\alias{as.numeric.Column}
\alias{as.integer.Column}
\alias{as.logical.Column}
\alias{as.POSIXct.Column}
\alias{as_datetime.Column}
\alias{as.array.Column}
\alias{as.list.Column}
\alias{is.numeric.Column}
\alias{is.logical.Column}
\alias{is.character.Column}
\alias{is.list.Column}
\title{Column Type Conversions}
\usage{
\method{as.character}{Column}(x, ...)

\method{as.numeric}{Column}(x, ...)

\method{as.integer}{Column}(x, ...)

\method{as.logical}{Column}(x, ...)

\method{as.POSIXct}{Column}(x, ...)

as_datetime.Column(x, ...)

\method{as.array}{Column}(x, ...)

\method{as.list}{Column}(x, ...)

\method{is.numeric}{Column}(x)

is.logical.Column(x)

is.character.Column(x)

is.list.Column(x)
}
\arguments{
\item{x}{a \code{Column} object}

\item{...}{addition arguments, currently unused}
}
\description{
these functions can be used to coerce \code{Column} objects
to other types
}
