% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/columns.R
\name{asc}
\alias{asc}
\alias{contains}
\alias{getField}
\alias{getItem}
\alias{isNaN}
\alias{isNull}
\alias{isNotNull}
\alias{like}
\alias{rlike}
\alias{columnfunctions}
\alias{asc,Column-method}
\alias{like,Column-method}
\alias{rlike,Column-method}
\alias{getField,Column-method}
\alias{getItem,Column-method}
\title{A set of operations working with SparkDataFrame columns}
\usage{
asc(x)

contains(x, ...)

getField(x, ...)

getItem(x, ...)

isNaN(x)

isNull(x)

isNotNull(x)

like(x, ...)

rlike(x, ...)

\S4method{asc}{Column}(x)

\S4method{like}{Column}(x, data)

\S4method{rlike}{Column}(x, data)

\S4method{getField}{Column}(x, data)

\S4method{getItem}{Column}(x, data)
}
\arguments{
\item{x}{a \code{Column} object}

\item{...}{additional arguments}

\item{data}{c \code{Column} or other object type used for comparison}
}
\description{
A set of operations working with SparkDataFrame columns
}
