% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{count}
\alias{count}
\title{#' @export
lag <- function(...) {
  UseMethod("lag")
}}
\usage{
count(...)
}
\description{
#' @importFrom dplyr lag
lag.default <- function(x, offset = 1, defaultValue = NULL) {
  dplyr::lag(x, offset = 1, defaultValue = NULL)
}
}
\details{
#' @importFrom dplyr lag
lag.Column <- function(x, offset = 1, defaultValue = NULL) {

col <- if (class(x) == "Column") {
    x@jc
  } else {
    x
  }

jc <- call_static("org.apache.spark.sql.functions", "lag", col, as.integer(offset), defaultValue)

new("Column", jc)
}
}
