% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{covariance}
\alias{covariance}
\alias{cov}
\title{Covariance on a Spark DataFrame}
\usage{
covariance(x, colName1, colName2)
}
\arguments{
\item{x}{a \code{spark_tbl}}

\item{colName1}{the name of the first column}

\item{colName2}{the name of the second column}
}
\value{
The covariance of the two columns.
}
\description{
Covariance on a Spark DataFrame
}
\details{
\code{cov}: When applied to spark_tbl, this calculates the sample
covariance of two numerical columns of \emph{one} spark_tbl.
}
\note{
cov since 1.6.0
}
\examples{

\dontrun{
cov(df, "mpg", "hp")
cov(df, df$mpg, df$hp)}
}
\seealso{
Other stat functions: 
\code{\link{approxQuantile}()},
\code{\link{corr}()},
\code{\link{crosstab}()},
\code{\link{freqItems}()},
\code{\link{sampleBy}()}
}
\concept{stat functions}
