% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_lambda}
\alias{create_lambda}
\title{Create o.a.s.sql.expressions.LambdaFunction corresponding
to transformation described by func.
Used by higher order functions.}
\usage{
create_lambda(fun)
}
\arguments{
\item{fun}{R \code{function} (unary, binary or ternary)
that transforms \code{Columns} into a \code{Column}}
}
\value{
JVM \code{LambdaFunction} object
}
\description{
Create o.a.s.sql.expressions.LambdaFunction corresponding
to transformation described by func.
Used by higher order functions.
}
