% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_tree.R
\docType{class}
\name{DecisionTreeClassificationModel-class}
\alias{DecisionTreeClassificationModel-class}
\title{S4 class that represents a DecisionTreeClassificationModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala DecisionTreeClassificationModel}
}
\description{
S4 class that represents a DecisionTreeClassificationModel
}
\note{
DecisionTreeClassificationModel since 2.3.0
}
