% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_supervised.R
\docType{class}
\name{DecisionTreeRegressionModel-class}
\alias{DecisionTreeRegressionModel-class}
\title{S4 class that represents a DecisionTreeRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala DecisionTreeRegressionModel}
}
\description{
S4 class that represents a DecisionTreeRegressionModel
}
\note{
DecisionTreeRegressionModel since 2.3.0
}
