% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{display}
\alias{display}
\title{Display a sample of a \code{spark_tbl}}
\usage{
display(x, n = NULL)
}
\arguments{
\item{x}{a \code{spark_tbl}}

\item{n}{numeric, the number of rows to collect}
}
\value{
a \code{spark_tbl}, invisibly
}
\description{
Display a sample of a \code{spark_tbl}
}
\examples{

spark_tbl(iris) \%>\% display
spark_tbl(mtcars) \%>\% display(15)
}
