% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{first}
\alias{first}
\alias{first,characterOrColumn-method}
\title{#' @details
#' \code{factorial}: Computes the factorial of the given value.
#'
#' @rdname column_math_functions
#' @aliases factorial factorial,Column-method
#' @note factorial since 1.5.0
setMethod("factorial",
          signature(x = "Column"),
          function(x) {
            jc <- call_static("org.apache.spark.sql.functions", "factorial", x@jc)
            new("Column", jc)
          })
first}
\usage{
first(x, ...)

\S4method{first}{characterOrColumn}(x, na.rm = FALSE)
}
\arguments{
\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\description{
Aggregate function: returns the first value in a group.
}
\details{
The function by default returns the first values it sees. It will return the first non-missing
value it sees when na.rm is set to true. If all values are missing, then NA is returned.
Note: the function is non-deterministic because its results depends on order of rows which
may be non-deterministic after a shuffle.
}
\note{
first(characterOrColumn) since 1.4.0
}
\examples{
\dontrun{
first(df$c)
first(df$c, TRUE)
}
}
\seealso{
Other aggregate functions: 
\code{\link{avg}()},
\code{\link{corr}()},
\code{\link{cov}()},
\code{\link{last}()}
}
\concept{aggregate functions}
