% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\docType{methods}
\name{fitted,BisectingKMeansModel-method}
\alias{fitted,BisectingKMeansModel-method}
\title{Get fitted result from a bisecting k-means model}
\usage{
\S4method{fitted}{BisectingKMeansModel}(object, method = c("centers",
  "classes"))
}
\arguments{
\item{method}{type of fitted results, \code{"centers"} for cluster centers
or \code{"classes"} for assigned classes.}
}
\value{
\code{fitted} returns a SparkDataFrame containing fitted values.
}
\description{
Get fitted result from a bisecting k-means model.
Note: A saved-loaded model does not support this method.
}
