% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\docType{methods}
\name{fitted,KMeansModel-method}
\alias{fitted,KMeansModel-method}
\title{Get fitted result from a k-means model}
\usage{
\S4method{fitted}{KMeansModel}(object, method = c("centers", "classes"))
}
\arguments{
\item{object}{a fitted k-means model.}

\item{method}{type of fitted results, \code{"centers"} for cluster centers
or \code{"classes"} for assigned classes.}

\item{...}{additional argument(s) passed to the method.}
}
\value{
\code{fitted} returns a SparkDataFrame containing fitted values.
}
\description{
Get fitted result from a k-means model, similarly to R's fitted().
Note: A saved-loaded model does not support this method.
}
