% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{as.float}
\alias{as.float}
\alias{as.float.Column}
\title{Float Vectors}
\usage{
as.float(x, ...)

as.float.Column(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Coerces objects of type \code{float}.
}
