% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate.R
\name{floor_date}
\alias{floor_date}
\title{Floor Date}
\usage{
floor_date(x, unit, ...)
}
\arguments{
\item{x}{a vector or \code{Column} of date-time objects}

\item{unit}{a character string specifying a time unit or a multiple of a
unit to be rounded to. Valid base units are second, minute, hour, day,
week, month, quarter and year. Non-Spark usage additionally supports
bimonth, season, and halfyear.}

\item{...}{lubridate version has \code{week_start} (not included in Spark).
When unit is weeks specify the reference day; 7 being Sunday.}
}
\description{
Floor Date
}
