% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\docType{class}
\name{GaussianMixtureModel-class}
\alias{GaussianMixtureModel-class}
\title{S4 class that represents a GaussianMixtureModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala GaussianMixtureModel}
}
\description{
S4 class that represents a GaussianMixtureModel
}
\note{
GaussianMixtureModel since 2.1.0
}
