% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_tree.R
\docType{class}
\name{GBTClassificationModel-class}
\alias{GBTClassificationModel-class}
\title{S4 class that represents a GBTClassificationModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala GBTClassificationModel}
}
\description{
S4 class that represents a GBTClassificationModel
}
\note{
GBTClassificationModel since 2.1.0
}
