% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_tree.R
\docType{class}
\name{GBTRegressionModel-class}
\alias{GBTRegressionModel-class}
\title{S4 class that represents a GBTRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala GBTRegressionModel}
}
\description{
S4 class that represents a GBTRegressionModel
}
\note{
GBTRegressionModel since 2.1.0
}
