% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_regression.R
\docType{class}
\name{GeneralizedLinearRegressionModel-class}
\alias{GeneralizedLinearRegressionModel-class}
\title{S4 class that represents a generalized linear model}
\arguments{
\item{jobj}{a Java object reference to the backing Scala GeneralizedLinearRegressionWrapper}
}
\description{
S4 class that represents a generalized linear model
}
\note{
GeneralizedLinearRegressionModel since 2.0.0
}
