% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hashCode}
\alias{hashCode}
\title{Compute the hashCode of an object}
\usage{
hashCode(key)
}
\arguments{
\item{key}{the object to be hashed}
}
\description{
Java-style function to compute the hashCode for the given
object. Returns an integer value.
}
\details{
This only works for integer, numeric and character types.
}
\examples{

\dontrun{

hashCode(1L) # 1
hashCode(1.0) # 1072693248
hashCode("1") # 49

}
}
