% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{is.character.Column}
\alias{is.character.Column}
\title{#' @export
setMethod("is.logical", signature(x = "Column"),
          function(x) {
            T
          })}
\usage{
is.character.Column(x)
}
\description{
#' @export
setMethod("is.logical", signature(x = "Column"),
          function(x) {
            T
          })
}
