% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/streaming.R
\name{isActive}
\alias{isActive}
\alias{isActive,StreamingQuery-method}
\title{isActive}
\usage{
isActive(x)

\S4method{isActive}{StreamingQuery}(x)
}
\arguments{
\item{x}{a StreamingQuery.}
}
\value{
TRUE if query is actively running, FALSE if stopped.
}
\description{
Returns TRUE if this query is actively running.
}
\note{
isActive(StreamingQuery) since 2.2.0

experimental
}
\examples{
\dontrun{ isActive(sq) }
}
\seealso{
Other StreamingQuery methods: 
\code{\link{awaitTermination}()},
\code{\link{lastProgress}()},
\code{\link{queryName}()},
\code{\link{status}()},
\code{\link{stopQuery}()}
}
\concept{StreamingQuery methods}
