% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_regression.R
\docType{class}
\name{IsotonicRegressionModel-class}
\alias{IsotonicRegressionModel-class}
\title{S4 class that represents an IsotonicRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala IsotonicRegressionModel}
}
\description{
S4 class that represents an IsotonicRegressionModel
}
\note{
IsotonicRegressionModel since 2.1.0
}
