% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{isStreaming}
\alias{isStreaming}
\alias{is_streaming}
\alias{isStreaming,SparkDataFrame-method}
\title{isStreaming}
\usage{
is_streaming(.data)
}
\arguments{
\item{.data}{a \code{spark_tbl}}
}
\value{
TRUE if this SparkDataFrame is from a streaming source
}
\description{
Returns TRUE if this SparkDataFrame contains one or more sources that continuously return data
as it arrives. A dataset that reads data from a streaming source must be executed as a
\code{StreamingQuery} using \code{write.stream}.
}
\note{
isStreaming since 2.2.0

experimental
}
\examples{
\dontrun{
spark_session()
df <- read_stream("socket", host = "localhost", port = 9999)
is_Streaming(df)
}
}
\seealso{
\link{read.stream} \link{write.stream}

Other SparkDataFrame functions: 
\code{\link{write.stream}}
}
\concept{SparkDataFrame functions}
