% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{call_method}
\alias{call_method}
\alias{call_static}
\alias{call_method_handled}
\alias{new_jobj}
\alias{call_static_handled}
\title{Call Java Classes and Methods}
\usage{
call_method(jobj, method, ...)

call_static(class, method, ...)

call_method_handled(jobj, method, ...)

new_jobj(class, ...)

call_static_handled(class, method, ...)
}
\arguments{
\item{jobj}{a valid \code{jobj} object}

\item{method}{character, the method being invoked}

\item{...}{any other arguments to be passed on}

\item{class}{character, the class being invoked}
}
\description{
A low-level interface to Java and Spark. These functions
are exported for programmers who want to interact with Spark directly.
See https://spark.apache.org/docs/latest/api/java/index.html?org/apache/spark/sql/Dataset.html.
}
\details{
The difference between \code{handled} calls and other calls is
that handled calls capture JVM exceptions more effectivly. They are used
in settings where the success of an operation is not assured. Practically
speaking, the handled calls are thin wrappers of the non-handled calls.
}
