% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-related.R
\docType{data}
\name{json_sample}
\alias{json_sample}
\title{Sample JSON Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.
}
\source{
\url{https://docs.databricks.com/data/data-sources/read-json.html}
}
\usage{
json_sample
}
\description{
A sample JSON file for examples and testing
}
\keyword{datasets}
