% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\docType{class}
\name{KMeansModel-class}
\alias{KMeansModel-class}
\title{S4 class that represents a KMeansModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala KMeansModel}
}
\description{
S4 class that represents a KMeansModel
}
\note{
KMeansModel since 2.0.0
}
