% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/streaming.R
\name{lastProgress}
\alias{lastProgress}
\alias{lastProgress,StreamingQuery-method}
\title{lastProgress}
\usage{
lastProgress(x)

\S4method{lastProgress}{StreamingQuery}(x)
}
\arguments{
\item{x}{a StreamingQuery.}
}
\description{
Prints the most recent progess update of this streaming query in JSON format.
}
\note{
lastProgress(StreamingQuery) since 2.2.0

experimental
}
\examples{
\dontrun{ lastProgress(sq) }
}
\seealso{
Other StreamingQuery methods: 
\code{\link{awaitTermination}()},
\code{\link{isActive}()},
\code{\link{queryName}()},
\code{\link{status}()},
\code{\link{stopQuery}()}
}
\concept{StreamingQuery methods}
