% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\docType{class}
\name{LDAModel-class}
\alias{LDAModel-class}
\title{S4 class that represents an LDAModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala LDAWrapper}
}
\description{
S4 class that represents an LDAModel
}
\note{
LDAModel since 2.1.0
}
