% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification.R
\docType{class}
\name{LinearSVCModel-class}
\alias{LinearSVCModel-class}
\title{S4 class that represents an LinearSVCModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala LinearSVCModel}
}
\description{
S4 class that represents an LinearSVCModel
}
\note{
LinearSVCModel since 2.2.0
}
