% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{lit}
\alias{lit}
\alias{as.Column}
\alias{as_Column}
\title{Create a Column of literal value}
\usage{
lit(.x)

as.Column(.x)

as_Column(.x)
}
\arguments{
\item{.x}{a literal value or a Column object. A literal value can be an
arbitrary value like a character, string, or boolean value.}
}
\value{
a Column object
}
\description{
Create a Column object out of character, numeric, or boolean
value
}
\examples{

# these do the same thing:
as.Column("derpin'")
sparkumn("all day")
lit("long")

}
