% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification.R
\docType{class}
\name{LogisticRegressionModel-class}
\alias{LogisticRegressionModel-class}
\title{S4 class that represents an LogisticRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala LogisticRegressionModel}
}
\description{
S4 class that represents an LogisticRegressionModel
}
\note{
LogisticRegressionModel since 2.1.0
}
