% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate.R
\name{lubridate-Column}
\alias{lubridate-Column}
\alias{as_date.Column}
\alias{year.Column}
\alias{month.Column}
\alias{day.Column}
\alias{wday.Column}
\alias{hour.Column}
\alias{minute.Column}
\alias{second.Column}
\alias{week.Column}
\alias{quarter.Column}
\title{Lubridate-style Column Functions}
\usage{
as_date.Column(x)

\method{year}{Column}(x, ...)

\method{month}{Column}(x, ...)

\method{day}{Column}(x, ...)

\method{wday}{Column}(x, ...)

\method{hour}{Column}(x, ...)

\method{minute}{Column}(x, ...)

\method{second}{Column}(x, ...)

week.Column(x, ...)

quarter.Column(x, ...)
}
\arguments{
\item{x}{a \code{Column} object}

\item{...}{other arguments}
}
\description{
Column functions in the style of the lubridate package
}
