% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\name{ml_gaussian_mixture}
\alias{ml_gaussian_mixture}
\alias{spark.gaussianMixture,SparkDataFrame,formula-method}
\title{Multivariate Gaussian Mixture Model (GMM)}
\usage{
ml_gaussian_mixture(data, formula, k = 2, maxIter = 100, tol = 0.01)
}
\arguments{
\item{data}{a SparkDataFrame for training.}

\item{formula}{a symbolic description of the model to be fitted. Currently only a few formula
operators are supported, including '~', '.', ':', '+', and '-'.
Note that the response variable of formula is empty in spark.gaussianMixture.}

\item{k}{number of independent Gaussians in the mixture model.}

\item{maxIter}{maximum iteration number.}

\item{tol}{the convergence tolerance.}

\item{...}{additional arguments passed to the method.}
}
\value{
\code{ml_gaussian_mixture} returns a fitted multivariate gaussian mixture model.
}
\description{
Fits multivariate gaussian mixture model against a SparkDataFrame, similarly to R's
mvnormalmixEM(). Users can call \code{summary} to print a summary of the fitted model,
\code{predict} to make predictions on new data, and \code{write.ml}/\code{read.ml}
to save/load fitted models.
}
\seealso{
mixtools: \url{https://cran.r-project.org/package=mixtools}
}
