% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\name{ml_kmeans}
\alias{ml_kmeans}
\alias{ml_kmeans,spark_tbl,formula-method}
\alias{summary,KMeansModel-method}
\alias{fitted,KMeansModel-method}
\alias{write_ml,KMeansModel,character-method}
\title{K-Means Clustering Model}
\usage{
ml_kmeans(
  data,
  formula,
  k = 2,
  maxIter = 20,
  initMode = c("k-means||", "random"),
  seed = NULL,
  initSteps = 2,
  tol = 1e-04
)

\S4method{summary}{KMeansModel}(object)

\S4method{fitted}{KMeansModel}(object, method = c("centers", "classes"))

\S4method{write_ml}{KMeansModel,character}(object, path, overwrite = FALSE)
}
\arguments{
\item{data}{a spark_tbl for training.}

\item{formula}{a symbolic description of the model to be fitted. Currently
only a few formula operators are supported, including '~',
'.', ':', '+', and '-'. Note that the response variable of
formula is empty in ml_kmeans.}

\item{k}{number of centers.}

\item{maxIter}{maximum iteration number.}

\item{initMode}{the initialization algorithm chosen to fit the model.}

\item{seed}{the random seed for cluster initialization.}

\item{initSteps}{the number of steps for the k-means|| initialization mode.
This is an advanced setting, the default of 2 is almost
always enough. Must be > 0.}

\item{tol}{convergence tolerance of iterations.}

\item{object}{a fitted k-means model.}

\item{method}{type of fitted results, \code{"centers"} for cluster centers
or \code{"classes"} for assigned classes.}

\item{path}{the directory where the model is saved.}

\item{overwrite}{overwrites or not if the output path already exists. Default is FALSE
which means throw exception if the output path exists.}

\item{...}{additional argument(s) passed to the method.}
}
\value{
\code{ml_kmeans} returns a fitted k-means model.

\code{summary} returns summary information of the fitted model, which is a list.
        The list includes the model's \code{k} (the configured number of cluster centers),
        \code{coefficients} (model cluster centers),
        \code{size} (number of data points in each cluster), \code{cluster}
        (cluster centers of the transformed data), {is.loaded} (whether the model is loaded
        from a saved file), and \code{clusterSize}
        (the actual number of cluster centers. When using initMode = "random",
        \code{clusterSize} may not equal to \code{k}).

\code{fitted} returns a spark_tbl containing fitted values.
}
\description{
Fits a k-means clustering model against a spark_tbl, similarly to R's
kmeans(). Users can call \code{summary} to print a summary of the fitted
model, \code{predict} to make predictions on new data, and \code{write_ml}/
\code{read_ml} to save/load fitted models.

Get fitted result from a k-means model, similarly to R's fitted().
Note: A saved-loaded model does not support this method.
}
\note{
summary(KMeansModel) since 2.0.0

write_ml(KMeansModel, character) since 2.0.0
}
\examples{
\dontrun{
spark_session()
t <- as.data.frame(Titanic)
df <- spark_tbl(t)
model <- ml_kmeans(df, Class ~ Survived, k = 4, initMode = "random")
summary(model)
}
}
