% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_supervised.R
\name{ml_random_forest}
\alias{ml_random_forest}
\alias{spark.randomForest,spark_tbl,formula-method}
\title{Random Forest Model for Regression and Classification}
\usage{
ml_random_forest(
  data,
  formula,
  type = c("regression", "classification"),
  maxDepth = 5,
  maxBins = 32,
  numTrees = 20,
  impurity = NULL,
  featureSubsetStrategy = "auto",
  seed = NULL,
  subsamplingRate = 1,
  minInstancesPerNode = 1,
  minInfoGain = 0,
  checkpointInterval = 10,
  maxMemoryInMB = 256,
  cacheNodeIds = FALSE,
  handleInvalid = c("error", "keep", "skip")
)
}
\arguments{
\item{data}{a spark_tbl for training.}

\item{formula}{a symbolic description of the model to be fitted. Currently only a few formula
operators are supported, including '~', ':', '+', and '-'.}

\item{type}{type of model, one of "regression" or "classification", to fit}

\item{maxDepth}{Maximum depth of the tree (>= 0).}

\item{maxBins}{Maximum number of bins used for discretizing continuous features and for choosing
how to split on features at each node. More bins give higher granularity. Must be
>= 2 and >= number of categories in any categorical feature.}

\item{numTrees}{Number of trees to train (>= 1).}

\item{impurity}{Criterion used for information gain calculation.
For regression, must be "variance". For classification, must be one of
"entropy" and "gini", default is "gini".}

\item{featureSubsetStrategy}{The number of features to consider for splits at each tree node.
Supported options: "auto" (choose automatically for task: If
                   numTrees == 1, set to "all." If numTrees > 1
                   (forest), set to "sqrt" for classification and
                   to "onethird" for regression),
                   "all" (use all features),
                   "onethird" (use 1/3 of the features),
                   "sqrt" (use sqrt(number of features)),
                   "log2" (use log2(number of features)),
                   "n": (when n is in the range (0, 1.0], use
                   n * number of features. When n is in the range
                   (1, number of features), use n features).
                   Default is "auto".}

\item{seed}{integer seed for random number generation.}

\item{subsamplingRate}{Fraction of the training data used for learning each decision tree, in
range (0, 1].}

\item{minInstancesPerNode}{Minimum number of instances each child must have after split.}

\item{minInfoGain}{Minimum information gain for a split to be considered at a tree node.}

\item{checkpointInterval}{Param for set checkpoint interval (>= 1) or disable checkpoint (-1).
Note: this setting will be ignored if the checkpoint directory is not
set.}

\item{maxMemoryInMB}{Maximum memory in MB allocated to histogram aggregation.}

\item{cacheNodeIds}{If FALSE, the algorithm will pass trees to executors to match instances with
nodes. If TRUE, the algorithm will cache node IDs for each instance. Caching
can speed up training of deeper trees. Users can set how often should the
cache be checkpointed or disable it by setting checkpointInterval.}

\item{handleInvalid}{How to handle invalid data (unseen labels or NULL values) in features and
label column of string type in classification model.
Supported options: "skip" (filter out rows with invalid data),
                   "error" (throw an error), "keep" (put invalid data in
                   a special additional bucket, at index numLabels). Default
                   is "error".}

\item{...}{additional arguments passed to the method.}
}
\value{
\code{ml_random_forest} returns a fitted Random Forest model.
}
\description{
\code{ml_random_forest} fits a Random Forest Regression model or Classification model on
a spark_tbl. Users can call \code{summary} to get a summary of the fitted Random Forest
model, \code{predict} to make predictions on new data, and \code{write_ml}/\code{read_ml} to
save/load fitted models.
For more details, see
\href{http://spark.apache.org/docs/latest/ml-classification-regression.html#random-forest-regression}{
Random Forest Regression} and
\href{http://spark.apache.org/docs/latest/ml-classification-regression.html#random-forest-classifier}{
Random Forest Classification}
}
\examples{
\dontrun{
# fit a Random Forest Regression Model
df <- spark_tbl(longley)
model <- ml_random_forest(df, Employed ~ ., type = "regression", maxDepth = 5, maxBins = 16)

# get the summary of the model
summary(model)
}
}
