% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_supervised.R
\name{ml_survival_regression}
\alias{ml_survival_regression}
\title{Accelerated Failure Time (AFT) Survival Regression Model}
\usage{
ml_survival_regression(data, formula, aggregationDepth = 2,
  stringIndexerOrderType = c("frequencyDesc", "frequencyAsc",
  "alphabetDesc", "alphabetAsc"))
}
\arguments{
\item{data}{a SparkDataFrame for training.}

\item{formula}{a symbolic description of the model to be fitted. Currently only a few formula
operators are supported, including '~', ':', '+', and '-'.
Note that operator '.' is not supported currently.}

\item{aggregationDepth}{The depth for treeAggregate (greater than or equal to 2). If the
dimensions of features or the number of partitions are large, this
param could be adjusted to a larger size. This is an expert parameter.
Default value should be good for most cases.}

\item{stringIndexerOrderType}{how to order categories of a string feature column. This is used to
decide the base level of a string feature as the last category
after ordering is dropped when encoding strings. Supported options
are "frequencyDesc", "frequencyAsc", "alphabetDesc", and
"alphabetAsc". The default value is "frequencyDesc". When the
ordering is set to "alphabetDesc", this drops the same category
as R when encoding strings.}

\item{...}{additional arguments passed to the method.}
}
\value{
\code{ml_survival_regression} returns a fitted AFT survival regression model.
}
\description{
\code{ml_survival_regression} fits an accelerated failure time (AFT) survival regression model on
a SparkDataFrame. Users can call \code{summary} to get a summary of the fitted AFT model,
\code{predict} to make predictions on new data, and \code{write.ml}/\code{read.ml} to
save/load fitted models.
}
\seealso{
survival: \url{https://cran.r-project.org/package=survival}
}
