% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_supervised.R
\docType{class}
\name{MultilayerPerceptronClassificationModel-class}
\alias{MultilayerPerceptronClassificationModel-class}
\title{S4 class that represents a MultilayerPerceptronClassificationModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala MultilayerPerceptronClassifierWrapper}
}
\description{
S4 class that represents a MultilayerPerceptronClassificationModel
}
\note{
MultilayerPerceptronClassificationModel since 2.1.0
}
