% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{n_partitions}
\alias{n_partitions}
\title{Get the Number of Partitions in a \code{spark_tbl}}
\usage{
n_partitions(.data)
}
\arguments{
\item{.data}{a spark_tbl}
}
\description{
Return the number of partitions
}
