% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification.R
\docType{class}
\name{NaiveBayesModel-class}
\alias{NaiveBayesModel-class}
\title{S4 class that represents a NaiveBayesModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala NaiveBayesWrapper}
}
\description{
S4 class that represents a NaiveBayesModel
}
\note{
NaiveBayesModel since 2.0.0
}
