% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\docType{methods}
\name{operations}
\alias{operations}
\alias{+,Column,numeric-method}
\alias{+,numeric,Column-method}
\alias{-,Column,numeric-method}
\alias{-,numeric,Column-method}
\alias{*,Column,numeric-method}
\alias{*,numeric,Column-method}
\alias{/,Column,numeric-method}
\alias{/,numeric,Column-method}
\alias{\%\%,Column,numeric-method}
\alias{\%\%,numeric,Column-method}
\alias{==,Column,numeric-method}
\alias{==,numeric,Column-method}
\alias{==,Column,character-method}
\alias{==,character,Column-method}
\alias{==,Column,logical-method}
\alias{==,logical,Column-method}
\alias{>,Column,numeric-method}
\alias{>,numeric,Column-method}
\alias{<,Column,numeric-method}
\alias{<,numeric,Column-method}
\alias{!=,Column,numeric-method}
\alias{!=,numeric,Column-method}
\alias{!=,Column,character-method}
\alias{!=,character,Column-method}
\alias{!=,Column,logical-method}
\alias{!=,logical,Column-method}
\alias{<=,Column,numeric-method}
\alias{<=,numeric,Column-method}
\alias{>=,Column,numeric-method}
\alias{>=,numeric,Column-method}
\title{Column Operations}
\usage{
\S4method{+}{Column,numeric}(e1, e2)

\S4method{+}{numeric,Column}(e1, e2)

\S4method{-}{Column,numeric}(e1, e2)

\S4method{-}{numeric,Column}(e1, e2)

\S4method{*}{Column,numeric}(e1, e2)

\S4method{*}{numeric,Column}(e1, e2)

\S4method{/}{Column,numeric}(e1, e2)

\S4method{/}{numeric,Column}(e1, e2)

\S4method{\%\%}{Column,numeric}(e1, e2)

\S4method{\%\%}{numeric,Column}(e1, e2)

\S4method{==}{Column,numeric}(e1, e2)

\S4method{==}{numeric,Column}(e1, e2)

\S4method{==}{Column,character}(e1, e2)

\S4method{==}{character,Column}(e1, e2)

\S4method{==}{Column,logical}(e1, e2)

\S4method{==}{logical,Column}(e1, e2)

\S4method{>}{Column,numeric}(e1, e2)

\S4method{>}{numeric,Column}(e1, e2)

\S4method{<}{Column,numeric}(e1, e2)

\S4method{<}{numeric,Column}(e1, e2)

\S4method{!=}{Column,numeric}(e1, e2)

\S4method{!=}{numeric,Column}(e1, e2)

\S4method{!=}{Column,character}(e1, e2)

\S4method{!=}{character,Column}(e1, e2)

\S4method{!=}{Column,logical}(e1, e2)

\S4method{!=}{logical,Column}(e1, e2)

\S4method{<=}{Column,numeric}(e1, e2)

\S4method{<=}{numeric,Column}(e1, e2)

\S4method{>=}{Column,numeric}(e1, e2)

\S4method{>=}{numeric,Column}(e1, e2)
}
\arguments{
\item{e1}{the LHS of the operation}

\item{e2}{the RHS of the operation}
}
\value{
an object of class \code{Column]}
}
\description{
Various Column operations
}
