% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarian}
\alias{ovarian}
\title{Ovarian Cancer Survival Data}
\format{
\describe{
  \item{futime}{survival or censoring time}
  \item{fustat}{censoring status}
  \item{age}{in years}
  \item{resid.ds}{residual disease present (1=no,2=yes)}
  \item{rx}{treatment group}
  \item{ecog.ps}{ECOG performance status (1 is better, see reference)}
  }
}
\source{
Terry Therneau
}
\usage{
ovarian
}
\description{
Survival in a randomised trial comparing two treatments for
ovarian cancer
}
\references{
Edmunson, J.H., Fleming, T.R., Decker, D.G., Malkasian, G.D.,
Jefferies, J.A., Webb, M.J., and Kvols, L.K., Different Chemotherapeutic
Sensitivities and Host Factors Affecting Prognosis in Advanced Ovarian
Carcinoma vs. Minimal Residual Disease. Cancer Treatment Reports,
63:241-47, 1979.
}
\keyword{datasets}
