% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{cache}
\alias{cache}
\alias{persist}
\alias{unpersist}
\alias{storage_level}
\title{Storage Functions}
\usage{
cache(.data)

persist(.data, newLevel)

unpersist(.data, blocking = TRUE)

storage_level(.data)
}
\arguments{
\item{.data}{}

\item{newLevel}{storage level chosen for the persistence. See available options in the details.}

\item{blocking}{boolean, whether to block until all blocks are deleted.}

\item{.data}{a \code{spark_tbl}}

\item{.data}{a \code{spark_tbl}}

\item{.data}{a \code{spark_tbl}}
}
\value{
a \code{spark_tbl}
}
\description{
Functions to manage the storage of \code{spark_tbl} objects.
\code{persist} is capable of various storage modes (see details), and
\code{cache} is shorthand for \code{"MEMORY_AND_DISK"} storage level.
}
\details{
For details of the supported storage levels, refer to
\href{http://spark.apache.org/docs/latest/rdd-programming-guide.html#rdd-persistence}{rdd-persistence}.
}
\examples{
\dontrun{
spark_session()
iris_tbl <- spark_tbl(iris)

storage_level(iris_tbl)

persist(iris_tbl, "MEMORY_AND_DISK")
storage_level(iris_tbl)

unpersist(iris_tbl)
storage_level(iris_tbl)
}
}
