% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\docType{class}
\name{PowerIterationClustering-class}
\alias{PowerIterationClustering-class}
\title{S4 class that represents a PowerIterationClustering}
\arguments{
\item{jobj}{a Java object reference to the backing Scala PowerIterationClustering}
}
\description{
S4 class that represents a PowerIterationClustering
}
\note{
PowerIterationClustering since 3.0.0
}
