% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_utils.R
\name{predict}
\alias{predict}
\title{Makes predictions from a MLlib model}
\description{
Makes predictions from a MLlib model. For more information, see the specific
MLlib model below.
}
\seealso{
\link{spark.als}, \link{spark.bisectingKmeans}, \link{spark.decisionTree},

\link{spark.gaussianMixture}, \link{spark.gbt},

\link{spark.glm}, \link{glm}, \link{spark.isoreg},

\link{spark.kmeans},

\link{spark.logit}, \link{spark.mlp}, \link{spark.naiveBayes},

\link{spark.randomForest}, \link{spark.survreg}, \link{spark.svmLinear}
}
