% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/streaming.R
\name{queryName}
\alias{queryName}
\alias{queryName,StreamingQuery-method}
\title{queryName}
\usage{
queryName(x)

\S4method{queryName}{StreamingQuery}(x)
}
\arguments{
\item{x}{a StreamingQuery.}
}
\value{
The name of the query, or NULL if not specified.
}
\description{
Returns the user-specified name of the query. This is specified in
\code{write.stream(df, queryName = "query")}. This name, if set, must be
unique across all active queries.
}
\note{
queryName(StreamingQuery) since 2.2.0

experimental
}
\examples{
\dontrun{ queryName(sq) }
}
\seealso{
\link{write.stream}

Other StreamingQuery methods: 
\code{\link{awaitTermination}()},
\code{\link{isActive}()},
\code{\link{lastProgress}()},
\code{\link{status}()},
\code{\link{stopQuery}()}
}
\concept{StreamingQuery methods}
