% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_tree.R
\docType{class}
\name{RandomForestClassificationModel-class}
\alias{RandomForestClassificationModel-class}
\title{S4 class that represents a RandomForestClassificationModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala RandomForestClassificationModel}
}
\description{
S4 class that represents a RandomForestClassificationModel
}
\note{
RandomForestClassificationModel since 2.1.0
}
