% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_tree.R
\docType{class}
\name{RandomForestRegressionModel-class}
\alias{RandomForestRegressionModel-class}
\title{S4 class that represents a RandomForestRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala RandomForestRegressionModel}
}
\description{
S4 class that represents a RandomForestRegressionModel
}
\note{
RandomForestRegressionModel since 2.1.0
}
