% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{repartition}
\alias{repartition}
\title{Repartition a \code{spark_tbl}}
\usage{
repartition(.data, n_partitions, partition_by)
}
\arguments{
\item{.data}{a data frame to be repartitioned}

\item{n_partitions}{integer, the target number of partitions}

\item{partition_by}{vector of column names used for partitioning, only
supported for Spark 2.0+}
}
\description{
Repartitions a spark_tbl. Optionally allos for vector of
columns to be used for partitioning.
}
\examples{
df <- spark_tbl(mtcars)

df \%>\% n_partitions() # 1

df_repartitioned <- df \%>\% repartition(5)
df \%>\% n_partitions() # 5

df_repartitioned <- df \%>\% repartition(5, c("cyl"))
}
