% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{repartition.spark_tbl}
\alias{repartition.spark_tbl}
\title{Repartition a \code{spark_tbl}}
\usage{
\method{repartition}{spark_tbl}(.data, n_partitions = NULL, partition_by = NULL)
}
\arguments{
\item{partition_by}{vector of column names used for partitioning, only
supported for Spark 2.0+}

\item{partitions}{number of partitions. Must be numeric.}
}
\description{
Repartitions a spark_tbl. Optionally allos for vector of
columns to be used for partitioning.
}
\examples{
df <- spark_tbl(mtcars)

df \%>\% n_partitions() # 1

df_repartitioned <- df \%>\% repartition(5)
df \%>\% n_partitions() # 5

df_repartitioned <- df \%>\% repartition(5, c("cyl"))
}
