% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{show}
\alias{show}
\title{Show a sample of a \code{spark_tbl}}
\usage{
show(x, n = NULL)
}
\arguments{
\item{x}{a \code{spark_tbl}}

\item{n}{numeric, the number of rows to collect}
}
\value{
a \code{spark_tbl}, invisibly
}
\description{
Show a sample of a \code{spark_tbl}
}
\examples{

spark_tbl(iris) \%>\% show
spark_tbl(mtcars) \%>\% show(15)
}
