% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_supervised.R
\name{spark.glm}
\alias{spark.glm}
\alias{ml_glm}
\alias{spark.glm,spark_tbl,formula-method}
\title{Generalized Linear Models}
\usage{
ml_glm(
  data,
  formula,
  family = "gaussian",
  tol = 1e-06,
  maxIter = 25,
  weightCol = NULL,
  regParam = 0,
  var.power = 0,
  link.power = 1 - var.power,
  stringIndexerOrderType = c("frequencyDesc", "frequencyAsc", "alphabetDesc",
    "alphabetAsc"),
  offsetCol = NULL
)
}
\arguments{
\item{data}{a spark_tbl for training.}

\item{formula}{a symbolic description of the model to be fitted. Currently only a few formula
operators are supported, including '~', '.', ':', '+', and '-'.}

\item{family}{a description of the error distribution and link function to be used in the model.
              This can be a character string naming a family function, a family function or
              the result of a call to a family function. Refer R family at
              \url{https://stat.ethz.ch/R-manual/R-devel/library/stats/html/family.html}.
              Currently these families are supported: \code{binomial}, \code{gaussian},
              \code{Gamma}, \code{poisson} and \code{tweedie}.

              Note that there are two ways to specify the tweedie family.
              \itemize{
               \item Set \code{family = "tweedie"} and specify the var.power and link.power;
               \item When package \code{statmod} is loaded, the tweedie family is specified
               using the family definition therein, i.e., \code{tweedie(var.power, link.power)}.
              }}

\item{tol}{positive convergence tolerance of iterations.}

\item{maxIter}{integer giving the maximal number of IRLS iterations.}

\item{weightCol}{the weight column name. If this is not set or \code{NULL}, we treat all instance
weights as 1.0.}

\item{regParam}{regularization parameter for L2 regularization.}

\item{var.power}{the power in the variance function of the Tweedie distribution which provides
the relationship between the variance and mean of the distribution. Only
applicable to the Tweedie family.}

\item{link.power}{the index in the power link function. Only applicable to the Tweedie family.}

\item{stringIndexerOrderType}{how to order categories of a string feature column. This is used to
decide the base level of a string feature as the last category
after ordering is dropped when encoding strings. Supported options
are "frequencyDesc", "frequencyAsc", "alphabetDesc", and
"alphabetAsc". The default value is "frequencyDesc". When the
ordering is set to "alphabetDesc", this drops the same category
as R when encoding strings.}

\item{offsetCol}{the offset column name. If this is not set or empty, we treat all instance
offsets as 0.0. The feature specified as offset has a constant coefficient of
1.0.}

\item{...}{additional arguments passed to the method.}
}
\value{
\code{spark.glm} returns a fitted generalized linear model.
}
\description{
Fits generalized linear model against a spark_tbl.
Users can call \code{summary} to print a summary of the fitted model, \code{predict} to make
predictions on new data, and \code{write_ml}/\code{read_ml} to save/load fitted models.
}
\examples{
\dontrun{
spark_session()
t <- as.data.frame(Titanic, stringsAsFactors = FALSE)
df <- spark_tbl(t)
model <- ml_glm(df, Freq ~ Sex + Age, family = "gaussian")
summary(model)
}
}
