% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_supervised.R
\name{spark.isoreg}
\alias{spark.isoreg}
\alias{ml_isotonic_regression}
\alias{spark.isoreg,SparkDataFrame,formula-method}
\title{Isotonic Regression Model}
\usage{
ml_isotonic_regression(
  data,
  formula,
  isotonic = TRUE,
  featureIndex = 0,
  weightCol = NULL
)
}
\arguments{
\item{data}{SparkDataFrame for training.}

\item{formula}{A symbolic description of the model to be fitted. Currently only a few formula
operators are supported, including '~', '.', ':', '+', and '-'.}

\item{isotonic}{Whether the output sequence should be isotonic/increasing (TRUE) or
antitonic/decreasing (FALSE).}

\item{featureIndex}{The index of the feature if \code{featuresCol} is a vector column
(default: 0), no effect otherwise.}

\item{weightCol}{The weight column name.}

\item{...}{additional arguments passed to the method.}
}
\value{
\code{spark.isoreg} returns a fitted Isotonic Regression model.
}
\description{
Fits an Isotonic Regression model against a SparkDataFrame, similarly to R's isoreg().
Users can print, make predictions on the produced model and save the model to the input path.
}
\examples{
\dontrun{
spark_session()
data <- list(list(7.0, 0.0), list(5.0, 1.0), list(3.0, 2.0),
        list(5.0, 3.0), list(1.0, 4.0))
df <- spark_tbl(data \%>\% setNames("label", "feature"))
model <- ml_isotonic_regression(df, label ~ feature, isotonic = FALSE)
# return model boundaries and prediction as lists
result <- summary(model, df)
}
}
