% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering.R
\name{spark.kmeans}
\alias{spark.kmeans}
\alias{ml_kmeans}
\alias{spark.kmeans,SparkDataFrame,formula-method}
\title{K-Means Clustering Model}
\usage{
ml_kmeans(data, formula, k = 2, maxIter = 20,
  initMode = c("k-means||", "random"), seed = NULL, initSteps = 2,
  tol = 1e-04)
}
\arguments{
\item{data}{a SparkDataFrame for training.}

\item{formula}{a symbolic description of the model to be fitted. Currently only a few formula
operators are supported, including '~', '.', ':', '+', and '-'.
Note that the response variable of formula is empty in spark.kmeans.}

\item{k}{number of centers.}

\item{maxIter}{maximum iteration number.}

\item{initMode}{the initialization algorithm chosen to fit the model.}

\item{seed}{the random seed for cluster initialization.}

\item{initSteps}{the number of steps for the k-means|| initialization mode.
This is an advanced setting, the default of 2 is almost always enough.
Must be > 0.}

\item{tol}{convergence tolerance of iterations.}

\item{...}{additional argument(s) passed to the method.}
}
\value{
\code{spark.kmeans} returns a fitted k-means model.
}
\description{
Fits a k-means clustering model against a SparkDataFrame, similarly to R's kmeans().
Users can call \code{summary} to print a summary of the fitted model, \code{predict} to make
predictions on new data, and \code{write.ml}/\code{read.ml} to save/load fitted models.
}
