% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{spark_class}
\alias{spark_class}
\title{Get Spark Class}
\usage{
spark_class(x, trunc)
}
\arguments{
\item{x}{a \code{spark_tbl} or \code{jobj}}

\item{trunc}{whether to return a truncated class name}
}
\value{
a character representing the spark object type
}
\description{
Get Spark Class
}
