% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{spark_lapply}
\alias{spark_lapply}
\title{Apply a Function over a List or Vector, Distribute operations in Spark}
\usage{
spark_lapply(.l, .f)
}
\arguments{
\item{.l}{a vector (atomic or list)}

\item{.f}{he function to be applied to each element of \code{.l}}
}
\value{
an in-memory \code{list} object
}
\description{
Run a function over a list of elements, distributing the
computations with Spark. Applies a function in a manner that is similar to
doParallel or lapply to elements of a list. The computations are distributed
using Spark. It is conceptually the same as the following
code: \code{lapply(list, func)}
}
\examples{

spark_session()
doubled <- spark_lapply(1:10, function(x) {2 * x})

# or using tidyverse style lamdas
doubled <- spark_lapply(1:10, ~ 2 * .)

}
