% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_read_json}
\alias{spark_read_json}
\title{Read a JSON file into a \code{spark_tbl}.}
\usage{
spark_read_json(path, ...)
}
\arguments{
\item{path}{string, the path to the file. Needs to be accessible from the cluster.}

\item{...}{named list, optional arguments to the reader}
}
\value{

}
\description{
Read a JSON file into a \code{spark_tbl}.
}
\examples{
TODO example of specifiying a schema and reading nested data
}
