% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_read_parquet}
\alias{spark_read_parquet}
\title{Read a parquet file into a \code{spark_tbl}.}
\usage{
spark_read_parquet(path, ...)
}
\arguments{
\item{path}{string, the path to the file. Needs to be accessible from the cluster.}

\item{...}{named list, optional arguments to the reader}
}
\value{
a \code{spark_tbl}
}
\description{
Read a parquet file into a \code{spark_tbl}.
}
\details{
Other options such as specifing a schema can be specified in the \code{...}
}
