% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_read_source}
\alias{spark_read_source}
\title{Read from a generic source into a \code{spark_tbl}}
\usage{
spark_read_source(
  path = NULL,
  source = NULL,
  schema = NULL,
  na.strings = "NA",
  ...
)
}
\arguments{
\item{path}{string, the path to the file. Needs to be accessible from the cluster.}

\item{source}{string, A data source capable of reading data.}

\item{schema}{FILL IN}

\item{na.strings}{string, the string value used to signify NA values.}

\item{...}{named list, optional arguments to the reader}
}
\value{
a \code{spark_tbl}
}
\description{
Read from a generic source into a \code{spark_tbl}
}
