% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_read_table}
\alias{spark_read_table}
\title{Read a Spark Managed Table}
\usage{
spark_read_table(table)
}
\arguments{
\item{table}{string, the name of the table to read}
}
\value{
a \code{spark_tbl}
}
\description{
a shortcut function to read a Spark-managed table directly
from the hive metastore without have to write any SQL. This is not a
feature in Spark's Scala or Python API.
}
\examples{
\dontrun{
spark_read_table("iris")
# same as
spark.sql("SELECT * FROM iris")
}
}
