% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{spark_session_stop}
\alias{spark_session_stop}
\title{Stop the Spark Session and Spark Context}
\usage{
spark_session_stop()
}
\description{
Stop the Spark Session and Spark Context.
}
\details{
Also terminates the backend this R session is connected to.
}
\examples{
spark_session_stop()
}
