% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{spark_sql}
\alias{spark_sql}
\title{Spark SQL}
\usage{
spark_sql(expr)
}
\arguments{
\item{expr}{a string, the SQL query}
}
\value{
a \code{spark_tbl}
}
\description{
Execute Spark SQL queries directly on the spark session
}
\examples{
\dontrun{
spark_tbl(iris) \%>\% register_temp_view("iris")
iris_preview <- spark_sql("SELECT * FROM iris LIMIT 10")
iris_preview \%>\% collect
}
}
