% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{spark_tbl}
\alias{spark_tbl}
\title{Create a \code{spark_tbl}}
\usage{
spark_tbl(x, ...)
}
\arguments{
\item{x}{object coercible to \code{spark_tbl}}

\item{...}{}
}
\value{
an object of class \code{spark_tbl}
}
\description{
Create a \code{spark_tbl}
}
\examples{
spark_tbl(iris)
spark_tbl(tibble(x = 1:10, y = 10:1))
spark_tbl(SparkR::as.DataFrame(iris))
}
