% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_write_csv}
\alias{spark_write_csv}
\title{Write a \code{spark_tbl} to CSV format}
\usage{
spark_write_csv(.data, path, mode = "error", partition_by = NULL, ...)
}
\arguments{
\item{.data}{a \code{spark_tbl}}

\item{path}{string, the path where the file is to be saved.}

\item{mode}{string, usually \code{"error"}, \code{"overwrite"}, or \code{"append"}/}

\item{partition_by}{string, column names to partition by on disk}

\item{...}{any other named options. See details below.}
}
\value{

}
\description{
Write a \code{spark_tbl} to a tabular (typically, comma-separated) file.
}
\details{
Many other options can be set using the \code{...}. Some popular
ones include \code{header = T} or \code{sep = ","}. A full list can be found
here: https://spark.apache.org/docs/latest/api/java/org/apache/spark/sql/DataFrameWriter.html#csv-java.lang.String-
}
