% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_write_delta}
\alias{spark_write_delta}
\title{Write a \code{spark_tbl} to Delta format}
\usage{
spark_write_delta(.data, path, mode = "error", partition_by = NULL, ...)
}
\arguments{
\item{.data}{a \code{spark_tbl}}

\item{path}{string, the path where the file is to be saved.}

\item{mode}{string, usually \code{"error"} (default), \code{"overwrite"},
\code{"append"}, or \code{"ignore"}}

\item{partition_by}{string, column names to partition by on disk}

\item{...}{any other named options. See details below.}
}
\description{
Write a \code{spark_tbl} to Delta
}
\details{
For text, a few additional options can be specified using \code{...}:
#' \describe{
  \item{compression}{(default null), compression codec to use when saving to
  file. This can be one of the known case-insensitive shorten names (none,
  bzip2, gzip, lz4, snappy and deflate).}
  \item{dateFormat}{(default yyyy-MM-dd), sets the string that indicates a
  date format. Custom date formats follow the formats at java.text.SimpleDateFormat.
  This applies to date type.}
  \item{timestampFormat}{(default yyyy-MM-dd'T'HH:mm:ss.SSSXXX), sets the
  string that indicates a timestamp format. Custom date formats follow the
  formats at java.text.SimpleDateFormat. This applies to timestamp type.}
  \item{encoding}{(by default it is not set), specifies encoding (charset)
  of saved json files. If it is not set, the UTF-8 charset will be used.}
  \item{lineSep}{(default \\n), defines the line separator that should be
  used for writing.}
}
}
