% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_write_insert}
\alias{spark_write_insert}
\title{Insert into a Spark Managed Table}
\usage{
spark_write_insert(.data, table, mode = "append")
}
\arguments{
\item{.data}{a \code{spark_tbl}}

\item{table}{string, the table name}

\item{mode}{string, usually \code{"append"} (default), \code{"overwrite"},
\code{"error"}, or \code{"ignore"}}

\item{partition_by}{string, column names to partition by}
}
\description{
Inserts the content of the \code{spark_tbl} into the specified
table. An R wrapper for Spark's \code{insertInto}.
}
\details{
Unlike \code{saveAsTable}, \code{insertInto} ignores the column
names and just uses position-based resolution. Watch out for column order!
}
