% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_write_jdbc}
\alias{spark_write_jdbc}
\title{Write to a JDBC table}
\usage{
spark_write_jdbc(
  .data,
  url,
  table = NULL,
  mode = "error",
  partition_by = NULL,
  driver = NULL,
  ...
)
}
\arguments{
\item{.data}{a \code{spark_tbl}}

\item{url}{string, the jdbc URL}

\item{table}{sting, the table name}

\item{mode}{string, either \code{"error"} (default), \code{"overwrite"},
or \code{"append"}.}

\item{partition_by}{string, the column to partition by}

\item{driver}{string, the driver class to use, e.g. \code{"org.postgresql.Driver"}}

\item{...}{additional connection options such as \code{user}, \code{password}, etc.}
}
\value{

}
\description{
Write to a JDBC table
}
