% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/streaming.R
\name{status}
\alias{status}
\alias{status,StreamingQuery-method}
\title{status}
\usage{
status(x)

\S4method{status}{StreamingQuery}(x)
}
\arguments{
\item{x}{a StreamingQuery.}
}
\description{
Prints the current status of the query in JSON format.
}
\note{
status(StreamingQuery) since 2.2.0

experimental
}
\examples{
\dontrun{ status(sq) }
}
\seealso{
Other StreamingQuery methods: 
\code{\link{awaitTermination}()},
\code{\link{isActive}()},
\code{\link{lastProgress}()},
\code{\link{queryName}()},
\code{\link{stopQuery}()}
}
\concept{StreamingQuery methods}
