% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\docType{class}
\name{StreamingQuery-class}
\alias{StreamingQuery-class}
\title{S4 class that represents a StreamingQuery}
\arguments{
\item{ssq}{A Java object reference to the backing Scala StreamingQuery}
}
\description{
StreamingQuery can be created by using read.stream() and write.stream()
}
\note{
StreamingQuery since 2.2.0

experimental
}
\seealso{
\link{read.stream}
}
