% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{structField}
\alias{structField}
\alias{structField.jobj}
\alias{structField.character}
\title{structField}
\usage{
structField(x, ...)

\method{structField}{jobj}(x, ...)

\method{structField}{character}(x, type, nullable = TRUE, ...)
}
\arguments{
\item{x}{the name of the field.}

\item{...}{additional argument(s) passed to the method.}

\item{type}{The data type of the field, see
https://spark.apache.org/docs/latest/sparkr.html#data-type-mapping-between-r-and-spark}

\item{nullable}{boolean, whether or not the field is nullable}
}
\value{

}
\description{
structField
}
\examples{
## Not run:
schema <- structType(
  structField("a", "integer"),
  structField("c", "string"),
  structField("avg", "double")
  )
## End(Not run)
}
