% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{structType}
\alias{structType}
\alias{structType.jobj}
\alias{structType.character}
\alias{structType.structField}
\title{structType}
\usage{
structType(x, ...)

\method{structType}{jobj}(x, ...)

\method{structType}{character}(x, ...)

\method{structType}{structField}(x, ...)
}
\arguments{
\item{x}{a structField object (created with the structField method).
Since Spark 2.3, this can be a DDL-formatted string, which is a comma
separated list of field definitions, e.g., "a INT, b STRING".}

\item{...}{additional structField objects}
}
\value{
an object of type \code{structType}
}
\description{
Create a structType object that contains the metadata for
a SparkDataFrame. Intended for use with read functions and apply functions.
}
\examples{
schema <- structType(
  structField("a", "integer"),
  structField("c", "string"),
  structField("avg", "double")
  )
}
