% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_supervised.R
\docType{methods}
\name{summary,MultilayerPerceptronClassificationModel-method}
\alias{summary,MultilayerPerceptronClassificationModel-method}
\title{importFrom("stats", "na.omit")
importFrom("utils", "tail")}
\usage{
\S4method{summary}{MultilayerPerceptronClassificationModel}(object)
}
\description{
importFrom("stats", "na.omit")
importFrom("utils", "tail")
}
