% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{as.timestamp}
\alias{as.timestamp}
\alias{as.timestamp.Column}
\title{Timestamp Vectors}
\usage{
as.timestamp(x, ...)

as.timestamp.Column(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Coerces objects of type \code{timestamp}.
}
