% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unresolved_named_lambda_var}
\alias{unresolved_named_lambda_var}
\title{Create o.a.s.sql.expressions.UnresolvedNamedLambdaVariable,
convert it to o.s.sql.Column and wrap with R Column.
Used by higher order functions.}
\usage{
unresolved_named_lambda_var(...)
}
\arguments{
\item{...}{character of length = 1
if length(...) > 1 then argument is interpreted as a nested
Column, for example \code{unresolved_named_lambda_var("a", "b", "c")}
yields unresolved \code{a.b.c}}
}
\value{
Column object wrapping JVM UnresolvedNamedLambdaVariable
}
\description{
Create o.a.s.sql.expressions.UnresolvedNamedLambdaVariable,
convert it to o.s.sql.Column and wrap with R Column.
Used by higher order functions.
}
