% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindowSpec.R
\docType{class}
\name{WindowSpec-class}
\alias{WindowSpec-class}
\title{S4 class that represents a WindowSpec}
\arguments{
\item{sws}{A Java object reference to the backing Scala WindowSpec}
}
\description{
WindowSpec can be created by using windowPartitionBy() or windowOrderBy()
}
\note{
WindowSpec since 2.0.0
}
\seealso{
\link{windowPartitionBy}, \link{windowOrderBy}
}
