% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_function_classes.R
\name{rank}
\alias{rank}
\alias{dense_rank}
\alias{min_rank}
\title{Rank}
\usage{
rank(x, ...)

dense_rank(x, ...)

min_rank(x, ...)
}
\arguments{
\item{x}{a vector or Column object}

\item{...}{additional arguments to \code{base::rank}}
}
\description{
Rank
}
