% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_function_classes.R
\name{rank}
\alias{rank}
\alias{dense_rank}
\alias{min_rank}
\title{#' @rdname column_window_functions
#' @name NULL
setGeneric("row_number", function(x = "missing") { standardGeneric("row_number") })}
\usage{
rank(x, ...)

dense_rank(x, ...)

min_rank(x, ...)
}
\description{
#' @export
setMethod("row_number", signature(x = "missing"),
          function(x) {
            jc <- SparkR:::callJStatic("org.apache.spark.sql.functions", "row_number")
            new("Column", jc)
          })
}
