% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/columns.R
\name{contains}
\alias{contains}
\alias{getField}
\alias{getItem}
\alias{isNaN}
\alias{isNull}
\alias{isNotNull}
\alias{like}
\alias{rlike}
\alias{columnfunctions}
\title{A set of operations working with SparkDataFrame columns}
\usage{
contains(x, ...)

getField(x, ...)

getItem(x, ...)

isNaN(x)

isNull(x)

isNotNull(x)

like(x, ...)

rlike(x, ...)
}
\arguments{
\item{x}{a Column object.}

\item{...}{additional argument(s).}
}
\description{
A set of operations working with Column objects
}
