% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{corr}
\alias{corr}
\alias{corr,Column-method}
\title{corr}
\usage{
corr(x, ...)

\S4method{corr}{Column}(x, col2)
}
\arguments{
\item{x}{a Column or a SparkDataFrame.}

\item{...}{additional argument(s). If \code{x} is a Column, a Column
should be provided. If \code{x} is a SparkDataFrame, two column names should
be provided.}

\item{col2}{a (second) Column.}
}
\description{
Computes the Pearson Correlation Coefficient for two Columns.
}
\note{
corr since 1.6.0
}
\examples{
\dontrun{
df <- createDataFrame(cbind(model = rownames(mtcars), mtcars))
head(select(df, corr(df$mpg, df$hp)))}
}
\seealso{
Other aggregate functions: \code{\link{avg}},
  \code{\link{cov}}, \code{\link{first}},
  \code{\link{last}}
}
\concept{aggregate functions}
