% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{covar_samp}
\alias{covar_samp}
\alias{covar_samp,characterOrColumn,characterOrColumn-method}
\alias{covar_pop}
\alias{covar_pop,characterOrColumn,characterOrColumn-method}
\title{Covariance}
\usage{
\S4method{covar_samp}{characterOrColumn,characterOrColumn}(col1, col2)

\S4method{covar_pop}{characterOrColumn,characterOrColumn}(col1, col2)
}
\arguments{
\item{col1}{the first Column.}

\item{col2}{the second Column.}
}
\description{
Compute the covariance between two expressions.
}
\details{
\code{covar_sample}: Alias for \code{cov}.

\code{covar_pop}: Computes the population covariance between two expressions.
}
\note{
covar_samp since 2.0.0

covar_pop since 2.0.0
}
