% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{firstItem}
\alias{firstItem}
\alias{firstItem,characterOrColumn-method}
\title{firstItem}
\usage{
firstItem(x, ...)

\S4method{firstItem}{characterOrColumn}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a character or \code{Column} object}

\item{...}{other arguments, currently unused}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\description{
Aggregate function: returns the first value in a group.
}
\details{
The function by default returns the first values it sees. It will return the first non-missing
value it sees when na.rm is set to true. If all values are missing, then NA is returned.
Note: the function is non-deterministic because its results depends on order of rows which
may be non-deterministic after a shuffle.
}
\note{
firstItem(characterOrColumn) since 1.4.0
}
\examples{
\dontrun{
firstItem(df$c)
firstItem(df$c, TRUE)
}
}
\seealso{
Other aggregate functions: 
\code{\link{approxCountDistinct}()},
\code{\link{avg}()},
\code{\link{corr}()},
\code{\link{lastItem}()}
}
\concept{aggregate functions}
