% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{get_spark_session}
\alias{get_spark_session}
\title{Get Spark Session}
\usage{
get_spark_session()
}
\value{
a \code{jobj} of the spark session
}
\description{
Get Spark Session
}
\examples{
get_spark_session()
}
