% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_tbl.R
\name{show}
\alias{show}
\title{show}
\usage{
show(.data, n = NULL)
}
\arguments{
\item{.data}{a \code{spark_tbl}}

\item{n}{numeric, the number of rows to collect}
}
\description{
If eager evaluation is enabled and the Spark object is a SparkDataFrame, evaluate the
SparkDataFrame and print top rows of the SparkDataFrame, otherwise, print the class
and type information of the Spark object.
}
