% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{spark_write_jdbc}
\alias{spark_write_jdbc}
\title{Write to a JDBC table}
\usage{
spark_write_jdbc(
  .data,
  url,
  table = NULL,
  mode = "error",
  partition_by = NULL,
  driver = NULL,
  ...
)
}
\arguments{
\item{.data}{a \code{spark_tbl}}

\item{url}{string, the jdbc URL}

\item{table}{sting, the table name}

\item{mode}{string, either \code{"error"} (default), \code{"overwrite"},
or \code{"append"}.}

\item{partition_by}{string, the column names to partition by}

\item{driver}{string, the driver class to use, e.g. \code{"org.postgresql.Driver"}}

\item{...}{additional connection options such as \code{user}, \code{password}, etc.}
}
\description{
Write to a JDBC table
}
\details{
connection properties can be set by other named arguments in the \code{...}
JDBC database connection arguments, a list of arbitrary string tag/value. Normally
at least a "user" and "password" property should be included. "batchsize" can be
used to control the number of rows per insert. "isolationLevel" can be one of
"NONE", "READ_COMMITTED", "READ_UNCOMMITTED", "REPEATABLE_READ", or "SERIALIZABLE",
corresponding to standard transaction isolation levels defined by JDBC's Connection
object, with default of "READ_UNCOMMITTED".
}
\examples{
spark_session_reset(sparkPackages = c("org.postgresql:postgresql:42.2.12"))

iris_tbl <- spark_tbl(iris)

iris_tbl \%>\%
  spark_write_jdbc(url = "jdbc:postgresql://localhost/tidyspark_test",
                   table = "iris_test",
                   partition_by = "Species",
                   mode = "overwrite",
                   user = "tidyspark_tester", password = "test4life",
                   driver = "org.postgresql.Driver")
}
