% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{StructField}
\alias{StructField}
\alias{StructField.jobj}
\alias{StructField.character}
\title{StructField}
\usage{
StructField(x, ...)

\method{StructField}{jobj}(x, ...)

\method{StructField}{character}(x, type, nullable = TRUE, ...)
}
\arguments{
\item{x}{the name of the field.}

\item{...}{additional argument(s) passed to the method.}

\item{type}{The data type of the field, see
https://spark.apache.org/docs/latest/sparkr.html#data-type-mapping-between-r-and-spark}

\item{nullable}{boolean, whether or not the field is nullable}
}
\value{
a StructField object
}
\description{
StructField
}
\examples{
## Not run:
schema <- StructType(
  StructField("a", "integer"),
  StructField("c", "string"),
  StructField("avg", "double")
  )
## End(Not run)
}
