% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{StructType}
\alias{StructType}
\alias{StructType.jobj}
\alias{StructType.character}
\alias{StructType.StructField}
\alias{StructType.StructType}
\title{StructType}
\usage{
StructType(x, ...)

\method{StructType}{jobj}(x, ...)

\method{StructType}{character}(x, ...)

\method{StructType}{StructField}(x, ...)

\method{StructType}{StructType}(x, ...)
}
\arguments{
\item{x}{a StructField object (created with the StructField method).
Since Spark 2.3, this can be a DDL-formatted string, which is a comma
separated list of field definitions, e.g., "a INT, b STRING".}

\item{...}{additional StructField objects}
}
\value{
an object of type \code{StructType}
}
\description{
Create a StructType object that contains the metadata for
a SparkDataFrame. Intended for use with read functions and apply functions.
}
\examples{
schema <- StructType(
  StructField("a", "integer"),
  StructField("c", "string"),
  StructField("avg", "double")
  )
}
