% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{as.timestamp}
\alias{as.timestamp}
\alias{as.timestamp.Column}
\alias{as.Date.Column}
\title{Timestamp and Date Vectors}
\usage{
as.timestamp(x, ...)

as.timestamp.Column(x, format = "yyyy-MM-dd HH:mm:ss", ...)

\method{as.Date}{Column}(x, format = "yyyy-MM-dd", ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}

\item{format}{a character string representing the incoming format. See
\href{here}{http://docs.oracle.com/javase/tutorial/i18n/format/simpleDateFormat.html}
for more information on possible usage.}
}
\description{
Coerces objects of type \code{timestamp}.
}
